@@ -7,12 +7,12 @@ description: A simple mobile app for tracking activities.
 # Both the version and the builder number may be overridden in flutter
 # build by specifying --build-name and --build-number, respectively.
 # Read more about versioning at semver.org.
-version: 1.0.6+20221024
+version: 1.1.0+20230117
 
 environment:
   # Note: Increasing min version here will enable null-safety checks (requires a TON of work to
   # resolve).
-  sdk: '>=2.11.0 <3.0.0'
+  sdk: '>=2.19.2 <3.0.0'
 
 dependencies:
   flutter:
@@ -37,9 +37,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^5.2.0               # https://pub.dev/packages/mockito
-  test: ^1.16.5                 # https://pub.dev/packages/test
-
+  build_runner: ^2.3.3          # https://pub.dev/packages/build_runner (Dart)
+  mockito: ^5.3.2               # https://pub.dev/packages/mockito (Dart)
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
