@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:mobile/res/dimen.dart';
 
 class PageAppBarStyle {
-  final String title;
-  final String subtitle;
-  final List<Widget> actions;
-  final Widget leading;
+  final String? title;
+  final String? subtitle;
+  final List<Widget>? actions;
+  final Widget? leading;
 
   PageAppBarStyle({
     this.title,
@@ -19,24 +19,23 @@ class PageAppBarStyle {
 
 class Page extends StatelessWidget {
   final Widget _child;
-  final PageAppBarStyle _appBarStyle;
+  final PageAppBarStyle? _appBarStyle;
 
   Page({
-    @required Widget child,
-    PageAppBarStyle appBarStyle,
-  }) : assert(child != null),
-       _child = child,
+    required Widget child,
+    PageAppBarStyle? appBarStyle,
+  }) : _child = child,
        _appBarStyle = appBarStyle;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: _appBarStyle == null ? null : AppBar(
-        title: _appBarStyle.subtitle == null
-            ? Text(_appBarStyle.title == null ? "" : _appBarStyle.title)
+        title: _appBarStyle!.subtitle == null
+            ? Text(_appBarStyle!.title == null ? "" : _appBarStyle!.title!)
             : _buildTitleWithSubtitle(context),
-        actions: _appBarStyle.actions,
-        leading: _appBarStyle.leading,
+        actions: _appBarStyle!.actions,
+        leading: _appBarStyle!.leading,
         elevation: 0,
       ),
       body: _child,
@@ -51,9 +50,9 @@ class Page extends StatelessWidget {
             ? CrossAxisAlignment.start
             : CrossAxisAlignment.center,
         children: <Widget>[
-          Text(_appBarStyle.title),
-          Text(_appBarStyle.subtitle,
-            style: Theme.of(context).textTheme.subtitle2.copyWith(
+          Text(_appBarStyle!.title ?? ""),
+          Text(_appBarStyle!.subtitle ?? "",
+            style: Theme.of(context).textTheme.subtitle2!.copyWith(
               color: Colors.white,
             ),
           ),
