@@ -1,13 +1,12 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mobile/utils/date_time_utils.dart';
 
 void main() {
   assertStatsDateRange({
-    @required DisplayDateRange dateRange,
-    @required DateTime now,
-    @required DateTime expectedStart,
-    DateTime expectedEnd,
+    required DisplayDateRange dateRange,
+    required DateTime now,
+    required DateTime expectedStart,
+    DateTime? expectedEnd,
   }) {
     DateRange range = dateRange.getValue(now);
     expect(range.startDate, equals(expectedStart));
